libelement: improvements to the C API & refactor
