+++ libelement/include/element/ast.h
@@ -5,8 +5,6 @@
 extern "C" {
 #endif
 
-#include <stdint.h>
-
 #include "element/common.h"
 #include "element/token.h"
 
@@ -39,24 +37,26 @@ typedef uint32_t element_ast_flags;
 // declaration flags
 static const element_ast_flags ELEMENT_AST_FLAG_DECL_INTRINSIC = (1U << 1);
 static const element_ast_flags ELEMENT_AST_FLAG_DECL_EMPTY_INPUT = (1U << 2);
-static const element_ast_flags ELEMENT_AST_FLAG_DECL_IMPLICIT_RETURN = (1U << 3);
 
 typedef struct element_ast_node element_ast_node;
 typedef struct element_ast element_ast;
 
-typedef struct element_parser_ctx element_parser_ctx;
+void element_ast_delete(element_ast** ast);
 
-element_result element_ast_get_type(const element_ast* node, element_ast_node_type* type);
-element_result element_ast_get_value_as_identifier(const element_ast* node, const char** value);
-element_result element_ast_get_value_as_literal(const element_ast* node, element_value* value);
+element_result element_ast_get_flags(const element_ast* ast, element_ast_flags* flags);
+element_result element_ast_get_nearest_token(const element_ast* ast, const element_token** token);
+element_result element_ast_get_type(const element_ast* ast, element_ast_node_type* type);
+element_result element_ast_get_value_as_identifier(const element_ast* ast, const char** value);
+element_result element_ast_get_value_as_literal(const element_ast* ast, element_value* value);
 element_result element_ast_get_parent(const element_ast* ast, element_ast** parent);
 element_result element_ast_get_child_count(const element_ast* ast, size_t* count);
-element_result element_ast_get_child(const element_ast* ast, const size_t index, element_ast** child);
-
-element_result element_ast_build(element_tokeniser_ctx* tctx, element_ast** ast);
-void element_ast_delete(element_ast* ast);
-
-element_result element_ast_to_string(const element_ast* ast, int depth, const element_ast* ast_to_mark, char* output_buffer, int output_buffer_size);
+element_result element_ast_get_child(const element_ast* ast, size_t index, element_ast** child);
+//may return ast if it is the root
+element_result element_ast_get_root(element_ast* ast, element_ast** root);
+//may return ast if it is the root
+element_result element_ast_get_root_const(const element_ast* ast, const element_ast** root);
+//ast_to_mark will output "HERE" beside it, pass null if unwanted
+element_result element_ast_to_string(const element_ast* ast, const element_ast* ast_to_mark, char* output_buffer, int output_buffer_size);
 
 #if defined(__cplusplus)
 }
