+++ libelement/include/element/element.h
@@ -1,0 +1,18 @@
+#if !defined(ELEMENT_ELEMENT_H)
+#define ELEMENT_ELEMENT_H
+
+#if defined(__cplusplus)
+extern "C" {
+#endif
+
+#include "ast.h"
+#include "common.h"
+#include "interpreter.h"
+#include "lmnt.h"
+#include "parser.h"
+#include "token.h"
+
+#if defined(__cplusplus)
+}
+#endif
+#endif
