--- libelement/notes.txt
@@ -1,13 +1,0 @@
-Remaining work
---------------
-Lambdas
-Constraints
-Type checking
-HOFs (ensure they are working correctly)
-Closures (ensure they are working correctly)
-
-Potential future work
----------------------
- - Consider whether we can allow for the input string alive for the lifetime of the interpreter to avoid string copies
- - Rework logging to tie concrete types to the particular error so that we can have compile-time type checking for structured errors
- - Rename to element to prevent liblibelement
