+++ libelement/src/token_internal.cpp
@@ -1,0 +1,486 @@
+#include "token_internal.hpp"
+
+//STD
+#include <cassert>
+#include "utf8.h"
+
+//LIBS
+#include <fmt/format.h>
+
+//SELF
+#include "interpreter_internal.hpp"
+
+//#define INCREMENT_TOKEN_LEN(s) { ++((s)->pos); ++((s)->character); ++((s)->cur_token.tok_len); }
+
+// #define UTF8_UNCHECKED
+#if defined(UTF8_UNCHECKED)
+#define UTF8_PEEK_NEXT(it, end) utf8::unchecked::peek_next(it)
+#define UTF8_NEXT(it, end) utf8::unchecked::next(it)
+#define UTF8_ADVANCE(it, n, end) utf8::unchecked::advance(it, n)
+#else
+#define UTF8_PEEK_NEXT(it, end) utf8::peek_next(it, end)
+#define UTF8_NEXT(it, end) utf8::next(it, end)
+#define UTF8_ADVANCE(it, n, end) utf8::advance(it, n, end)
+#endif
+
+#define PRINTCASE(a) \
+    case a:          \
+        c = #a;      \
+        break;
+
+static bool isid_alpha(uint32_t c) { return element_isalpha(c) || c == '_' || (c >= 0x00F0 && c <= 0xFFFF); }
+static bool isid_alnum(uint32_t c) { return element_isalnum(c) || (c >= 0x00F0 && c <= 0xFFFF); }
+
+void advance_to_end_of_line(std::string::const_iterator& it, const std::string::const_iterator& end)
+{
+    try
+    {
+        do
+        {
+            UTF8_NEXT(it, end);
+        } while (it != end && !element_iseol(UTF8_PEEK_NEXT(it, end)));
+    }
+    catch (...)
+    {
+        //exceptions are thrown for just about any utf issue, ignore them
+    }
+}
+
+element_tokeniser_ctx::element_tokeniser_ctx()
+{
+    tokens.reserve(512);
+    cur_token.type = ELEMENT_TOK_NONE;
+    cur_token.post_pos = -1;
+    cur_token.post_len = -1;
+    cur_token.tok_pos = -1;
+    cur_token.tok_len = -1;
+    cur_token.pre_pos = -1;
+    cur_token.pre_len = -1;
+}
+
+void element_tokeniser_ctx::reset_token()
+{
+    if (cur_token.type != ELEMENT_TOK_NONE)
+    {
+        // save current token
+        tokens.push_back(cur_token);
+        cur_token.pre_pos = -1;
+        cur_token.pre_len = 0;
+    }
+
+    cur_token.type = ELEMENT_TOK_NONE;
+    cur_token.tok_pos = -1;
+    cur_token.tok_len = 0;
+    cur_token.post_pos = -1;
+    cur_token.post_len = 0;
+
+    cur_token.line = -1;
+    cur_token.character = -1;
+    cur_token.line_start_position = -1;
+    cur_token.file_name = raw_filename;
+}
+
+element_result element_tokeniser_ctx::run(const char* cinput, const char* cfilename)
+{
+    raw_filename = cfilename;
+    filename = cfilename;
+    input = cinput;
+    pos = 0;
+    line = 1;
+    character = 1;
+    reset_token();
+
+    try
+    {
+        auto it = input.begin();
+        auto end = input.end();
+        uint32_t c;
+        while (it != end)
+        {
+            c = UTF8_PEEK_NEXT(it, end);
+            switch (c)
+            {
+            case '.':
+                add_token(ELEMENT_TOK_DOT, 1);
+                UTF8_NEXT(it, end);
+                break;
+            case '(':
+                add_token(ELEMENT_TOK_BRACKETL, 1);
+                UTF8_NEXT(it, end);
+                break;
+            case ')':
+                add_token(ELEMENT_TOK_BRACKETR, 1);
+                UTF8_NEXT(it, end);
+                break;
+            case ',':
+                add_token(ELEMENT_TOK_COMMA, 1);
+                UTF8_NEXT(it, end);
+                break;
+            case ':':
+                add_token(ELEMENT_TOK_COLON, 1);
+                UTF8_NEXT(it, end);
+                break;
+            case '{':
+                add_token(ELEMENT_TOK_BRACEL, 1);
+                UTF8_NEXT(it, end);
+                break;
+            case '}':
+                add_token(ELEMENT_TOK_BRACER, 1);
+                UTF8_NEXT(it, end);
+                break;
+            case '=':
+                add_token(ELEMENT_TOK_EQUALS, 1);
+                UTF8_NEXT(it, end);
+                break;
+            case '#':
+                tokenise_comment(it, end);
+                break;
+            case '_':
+            {
+                //in case there is nothing after this, add the token first first, and remove later if it's an identifier
+                add_token(ELEMENT_TOK_UNDERSCORE, 1);
+
+                const auto next = UTF8_PEEK_NEXT(it + 1, end);
+                if (isid_alpha(next))
+                {
+                    //remove token
+                    tokens.pop_back();
+                    pos -= 1;
+                    character -= 1;
+
+                    const auto begin_it = it;
+                    const auto result = tokenise_identifier(it, end);
+                    if (result != ELEMENT_OK)
+                    {
+                        log(result,
+                                       fmt::format("Failed to parse identifier '{}'",
+                                                   std::string(begin_it, it)));
+                        return result;
+                    }
+                }
+                else
+                {
+                    UTF8_NEXT(it, end);
+                }
+                break;
+            }
+            default:
+            {
+                if (isid_alpha(c))
+                {
+                    const auto begin_it = it;
+                    const auto result = tokenise_identifier(it, end);
+                    if (result != ELEMENT_OK)
+                    {
+                        log(result,
+                                       fmt::format("Failed to parse identifier '{}'",
+                                                   std::string(begin_it, it)));
+                        return result;
+                    }
+                }
+                else if (element_isdigit(c) || c == '-' || c == '+')
+                {
+                    const auto begin_it = it;
+                    const auto result = tokenise_number(it, end);
+                    if (result != ELEMENT_OK)
+                    {
+                        log(result,
+                                       fmt::format("Failed to parse number '{}'",
+                                                   std::string(begin_it, it)));
+                        return result;
+                    }
+                }
+                else if (element_iseol(c))
+                {
+                    ++line;
+                    character = 1;
+                    pos += 1;
+                    line_start_position = pos;
+                    line_number_to_line_pos.push_back(line_start_position);
+                    reset_token();
+                    UTF8_NEXT(it, end);
+                }
+                else if (element_isspace(c))
+                {
+                    //just eat it!
+                    pos += 1;
+                    character += 1;
+                    UTF8_NEXT(it, end);
+                }
+                else
+                {
+                    const auto begin_it = it;
+                    UTF8_NEXT(it, end);
+                    std::string source_line;
+                    source_line.resize(1024);
+                    assert(pos - line_start_position >= 0);
+                    assert(pos >= 0);
+                    memcpy(source_line.data(), cinput + line_start_position, static_cast<std::size_t>(pos) - line_start_position);
+                    log(ELEMENT_ERROR_PARSE,
+                                   fmt::format("Encountered invalid character '{}' in file {} on line {} character {}\n{}",
+                                               std::string(begin_it, it), cfilename, line, character, source_line));
+                    return ELEMENT_ERROR_PARSE;
+                }
+            }
+            }
+        }
+        add_token(ELEMENT_TOK_EOF, 0);
+        return ELEMENT_OK;
+    }
+    catch (const std::exception& e)
+    {
+        log(ELEMENT_ERROR_EXCEPTION,
+                       fmt::format("Exception occured: {}", e.what()));
+
+        return ELEMENT_ERROR_EXCEPTION;
+    }
+    catch (...) //potentially EOF when last source character is UTF?
+    {
+        log(ELEMENT_ERROR_EXCEPTION,
+                       fmt::format("Exception occured"));
+
+        return ELEMENT_ERROR_EXCEPTION;
+    }
+}
+
+std::string element_tokeniser_ctx::text(const element_token* t) const
+{
+    return input.substr(t->tok_pos, t->tok_len);
+}
+
+element_result element_tokeniser_ctx::tokenise_number(std::string::iterator& it, const std::string::iterator& end)
+{
+    //TODO: Go through this in detail, also, waaaaaaaaaaaaaaaaaaaaay too long for parsing a number
+    assert(cur_token.type == ELEMENT_TOK_NONE);
+    cur_token.type = ELEMENT_TOK_NUMBER;
+    cur_token.tok_pos = pos;
+    cur_token.line = line;
+    cur_token.line_start_position = line_start_position;
+    cur_token.character = character;
+
+    const auto it_begin = it;
+    uint32_t c = UTF8_PEEK_NEXT(it, end);
+    if (c == '-' || c == '+')
+    {
+        UTF8_NEXT(it, end);
+        c = UTF8_PEEK_NEXT(it, end);
+    }
+    assert(element_isdigit(c));
+
+    while (it != end && element_isdigit(UTF8_PEEK_NEXT(it, end)))
+        c = UTF8_NEXT(it, end);
+
+    if (it == end)
+    {
+        const size_t len = std::distance(it_begin, it);
+        pos += (int)len;
+        character += (int)len;
+        cur_token.tok_len += (int)len;
+        reset_token();
+        return ELEMENT_OK;
+    }
+
+    c = UTF8_PEEK_NEXT(it, end);
+    if (c == '.')
+    {
+        auto it_next = it;
+        UTF8_NEXT(it_next, end);
+
+        auto c_next = UTF8_PEEK_NEXT(it_next, end);
+        if (element_isdigit(c_next))
+        {
+            // number
+            UTF8_NEXT(it, end);
+            while (it != end && element_isdigit(UTF8_PEEK_NEXT(it, end)))
+                c = UTF8_NEXT(it, end);
+        }
+        else if (element_isalpha(c_next))
+        {
+            // indexing into a literal, do nothing and let the cleanup back out
+        }
+        else
+        {
+            const auto it_rhs_start = it_next;
+            UTF8_NEXT(it_next, end);
+            log(ELEMENT_ERROR_BAD_INDEX_INTO_NUMBER,
+                       fmt::format("Found {} which was thought to be a number being indexed, "
+                                   "but encountered invalid character '{}' on the right hand side of '.'",
+                                   std::string(it_begin, it), std::string(it_rhs_start, it_next)));
+
+            return ELEMENT_ERROR_BAD_INDEX_INTO_NUMBER;
+        }
+    }
+
+    if (it != end)
+    {
+        c = UTF8_PEEK_NEXT(it, end);
+        if (c == 'e' || c == 'E')
+        {
+            auto it_prev_character = it;
+            UTF8_NEXT(it, end);
+            c = UTF8_PEEK_NEXT(it, end);
+
+            if (c == '-' || c == '+')
+            {
+                it_prev_character = it;
+                UTF8_NEXT(it, end);
+                c = UTF8_PEEK_NEXT(it, end);
+            }
+
+            if (!element_isdigit(c))
+            {
+                log(ELEMENT_ERROR_BAD_NUMBER_EXPONENT,
+                           fmt::format("Found {} which was thought to be a number in scientific notation, "
+                                       "but encountered invalid character '{}' instead of the exponent number",
+                                       std::string(it_begin, it), std::string(it_prev_character, it)));
+                return ELEMENT_ERROR_BAD_NUMBER_EXPONENT;
+            }
+
+            while (it != end && element_isdigit(UTF8_PEEK_NEXT(it, end)))
+                c = UTF8_NEXT(it, end);
+        }
+    }
+
+    // determine length in bytes
+    const size_t len = std::distance(it_begin, it);
+    pos += (int)len;
+    character += (int)len;
+    cur_token.tok_len += (int)len;
+    reset_token();
+    return ELEMENT_OK;
+}
+
+element_result element_tokeniser_ctx::tokenise_comment(std::string::iterator& it, const std::string::iterator& end)
+{
+    //TODO: Go through this in detail
+    if (cur_token.post_pos < 0)
+        cur_token.post_pos = pos;
+
+    //consume all characters until end of file or end of line
+    const auto it_before = it;
+    try
+    {
+        while (it != end && !element_iseol(UTF8_PEEK_NEXT(it, end)))
+        { //will throw at EOF
+            UTF8_NEXT(it, end);
+        }
+    }
+    catch (...)
+    {
+        //we probably hit EOF. not the nicest way of handling this issue
+    }
+
+    // calculate correct length
+    const size_t len = std::distance(it_before, it);
+    cur_token.post_len += (int)len;
+    pos += (int)len;
+    character += (int)len;
+
+    return ELEMENT_OK;
+}
+
+element_result element_tokeniser_ctx::tokenise_identifier(std::string::iterator& it, const std::string::iterator& end)
+{
+    //TODO: Go through this in detail
+    assert(cur_token.type == ELEMENT_TOK_NONE);
+    cur_token.type = ELEMENT_TOK_IDENTIFIER;
+    cur_token.tok_pos = pos;
+    cur_token.line = line;
+    cur_token.line_start_position = line_start_position;
+    cur_token.character = character;
+    const auto it_begin = it;
+    uint32_t c = UTF8_PEEK_NEXT(it, end);
+    if (c == '_')
+    {
+        c = UTF8_NEXT(it, end);
+    }
+
+    assert(isid_alpha(c));
+
+    while (it != end
+           && (isid_alnum(UTF8_PEEK_NEXT(it, end)) || UTF8_PEEK_NEXT(it, end) == '_'))
+    {
+        c = UTF8_NEXT(it, end);
+    }
+
+    // determine length in bytes
+    const size_t len = std::distance(it_begin, it);
+    pos += (int)len;
+    character += (int)len;
+    cur_token.tok_len += (int)len;
+    reset_token();
+    return ELEMENT_OK;
+}
+
+void element_tokeniser_ctx::add_token(element_token_type t, int n)
+{
+    assert(cur_token.type == ELEMENT_TOK_NONE);
+    cur_token.type = t;
+    cur_token.tok_pos = pos;
+    cur_token.tok_len = n;
+    cur_token.line = line;
+    cur_token.line_start_position = line_start_position;
+    cur_token.character = character;
+    pos += n;
+    character += n;
+    reset_token();
+}
+
+std::string element_tokeniser_ctx::text_on_line(int line) const
+{
+    //lines start at 1, arrays at 0
+    line--;
+
+    if (line < 0 || line > line_number_to_line_pos.size())
+        return "invalid line";
+
+    const auto start_pos = line_number_to_line_pos[line];
+    const auto start_it = input.begin() + start_pos;
+    auto end_it = start_it;
+    advance_to_end_of_line(end_it, input.end());
+
+    return std::string(start_it, end_it);
+}
+
+void element_tokeniser_ctx::clear()
+{
+    tokens.clear();
+    filename.clear();
+    input.clear();
+    line = 1;
+    line_start_position = 0;
+    character = 1;
+    pos = 0;
+    reset_token();
+}
+
+void element_tokeniser_ctx::log(element_result message_code, const std::string& message) const
+{
+    if (logger == nullptr)
+        return;
+
+    logger->log(*this, message_code, message);
+}
+
+void element_tokeniser_ctx::log(element_result message_code, const std::string& message, int length, element_log_message* related_message) const
+{
+    if (logger == nullptr)
+        return;
+
+    logger->log(*this, message_code, message, length, related_message);
+}
+
+void element_tokeniser_ctx::log(const std::string& message) const
+{
+    if (logger == nullptr)
+        return;
+
+    logger->log(message, element_stage::ELEMENT_STAGE_MISC);
+}
+
+void element_tokeniser_ctx::set_log_callback(LogCallback callback, void* user_data)
+{
+    logger = std::make_shared<element_log_ctx>();
+    logger->callback = callback;
+    logger->user_data = user_data;
+}
