+++ libelement/src/token_internal.hpp
@@ -1,16 +1,39 @@
 #pragma once
 
-#include <cassert>
+//STD
 #include <string>
 #include <vector>
 #include <memory>
 #include <unordered_map>
 
+//SELF
 #include "element/token.h"
 #include "common_internal.hpp"
 
 struct element_tokeniser_ctx
 {
+public:
+    element_tokeniser_ctx();
+
+    element_result run(const char* cinput, const char* cfilename);
+    void clear();
+    void reset_token();
+    void add_token(element_token_type t, int n);
+
+    // literal ::= [-+]? [0-9]+ ('.' [0-9]*)? ([eE] [-+]? [0-9]+)?
+    element_result tokenise_number(std::string::iterator& it, const std::string::iterator& end);
+    element_result tokenise_comment(std::string::iterator& it, const std::string::iterator& end);
+    // identifier ::= '_'? [a-zA-Z\u00F0-\uFFFF] [_a-zA-Z0-9\u00F0-\uFFFF]*
+    element_result tokenise_identifier(std::string::iterator& it, const std::string::iterator& end);
+
+    [[nodiscard]] std::string text(const element_token* t) const;
+    [[nodiscard]] std::string text_on_line(int line) const;
+
+    void log(element_result message_code, const std::string& message) const;
+    void log(element_result message_code, const std::string& message, int length, element_log_message* related_message) const;
+    void log(const std::string& message) const;
+    void set_log_callback(LogCallback callback, void* user_data);
+
     std::shared_ptr<element_log_ctx> logger = nullptr;
     const char* raw_filename = nullptr;
     std::string filename;
@@ -22,56 +45,6 @@ struct element_tokeniser_ctx
     element_token cur_token;
     std::vector<element_token> tokens;
     std::vector<int> line_number_to_line_pos{ 0 };
-
-    std::string text(const element_token* t) const
-    {
-        return input.substr(t->tok_pos, t->tok_len);
-    }
-
-    std::string text_on_line(int line) const;
-
-    void log(int message_code, const std::string& message) const
-    {
-        if (logger == nullptr)
-            return;
-
-        logger->log(*this, message_code, message);
-    }
-
-    void log(int message_code, const std::string& message, int length, element_log_message* related_message) const
-    {
-        if (logger == nullptr)
-            return;
-
-        logger->log(*this, message_code, message, length, related_message);
-    }
-
-    void log(const std::string& message) const
-    {
-        if (logger == nullptr)
-            return;
-
-        logger->log(message, message_stage::ELEMENT_STAGE_MISC);
-    }
-
-    void set_log_callback(LogCallback callback, void* user_data)
-    {
-        logger = std::make_shared<element_log_ctx>();
-        logger->callback = callback;
-        logger->user_data = user_data;
-    }
-
-    element_tokeniser_ctx()
-    {
-        tokens.reserve(512);
-        cur_token.type = ELEMENT_TOK_NONE;
-        cur_token.post_pos = -1;
-        cur_token.post_len = -1;
-        cur_token.tok_pos = -1;
-        cur_token.tok_len = -1;
-        cur_token.pre_pos = -1;
-        cur_token.pre_len = -1;
-    }
 };
 
 // replacements for C stdlib functions which rely on slow locale stuff
