+++ libelement/test/tokeniser.test.cpp
@@ -1,15 +1,17 @@
-#include <catch2/catch.hpp>
-
-#include <element/token.h>
-#include "../src/token_internal.hpp"
-
 //STD
 #include <array>
+
+//LIBS
+#include <catch2/catch.hpp>
+
+//SELF
+#include "element/token.h"
+#include "token_internal.hpp"
 
 void print_token(const element_tokeniser_ctx* context, const element_token* nearest_token)
 {
     std::array<char, 2048> output_buffer{};
-    element_tokens_to_string(context, nearest_token, output_buffer.data(), output_buffer.size());
+    element_tokeniser_to_string(context, nearest_token, output_buffer.data(), output_buffer.size());
     printf("%s", output_buffer.data());
     UNSCOPED_INFO(output_buffer.data());
 }
